# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.addSecret:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  handlers.rbacGetResponse:
    properties:
      enabled:
        type: boolean
    type: object
  handlers.rbacPutRequest:
    properties:
      enabled:
        type: boolean
    type: object
...
host: gaia-pipeline.io
info:
  contact:
    name: API Support
    url: https://github.com/gaia-pipeline/gaia
  description: This is the API that the Gaia Admin UI uses.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/gaia-pipeline/gaia/blob/master/LICENSE
  title: Gaia API
  version: "1.0"
paths:
...
  /secret:
    post:
      consumes:
      - application/json
      description: Creates a secret.
      parameters:
      - description: The secret to create
        in: body
        name: secret
        required: true
        schema:
          $ref: '#/definitions/handlers.addSecret'
      produces:
      - application/json
      responses:
        "201":
          description: secret successfully set
          schema:
            type: string
        "400":
          description: Error binding or key is reserved.
          schema:
            type: string
        "500":
          description: Cannot get or load secrets
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a secret.
      tags:
      - secrets
  /secret/:key:
    delete:
      description: Removes a secret from the vault.
      parameters:
      - description: Key
        in: body
        name: key
        required: true
        schema:
          type: string
      produces:
      - text/plain
      responses:
        "200":
          description: secret successfully deleted
          schema:
            type: string
        "400":
          description: key is reserved and cannot be deleted
          schema:
            type: string
        "500":
          description: Cannot get or load secrets
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Removes a secret from the vault..
      tags:
      - secrets
  /secrets:
    get:
      description: Retrieves all secrets from the vault.
      produces:
      - application/json
      responses:
        "200":
          description: Secrets
          schema:
            items:
              $ref: '#/definitions/handlers.addSecret'
            type: array
        "500":
          description: Cannot get or load secrets
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: List all secrets.
      tags:
      - secrets
...
  /settings/rbac:
    get:
      description: Get the given RBAC settings.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.rbacGetResponse'
        "500":
          description: Something went wrong while saving or retrieving rbac settings.
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get RBAC settings
      tags:
      - settings
    put:
      consumes:
      - application/json
      description: Save the given RBAC settings.
      parameters:
      - description: RBAC setting details.
        in: body
        name: RbacPutRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.rbacPutRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: Settings have been updated.
          schema:
            type: string
        "400":
          description: Invalid body.
          schema:
            type: string
        "500":
          description: Something went wrong while saving or retrieving rbac settings.
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Put RBAC settings
      tags:
      - settings
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"